requirement.txt
Package                 Version     Editable project location
----------------------- ----------- -------------------------------------------
annoy                   1.17.0
antlr4-python3-runtime  4.8
appdirs                 1.4.4
backcall                0.2.0
black                   21.4b2
bpy                     2.91a0
click                   8.1.3
cloudpickle             2.1.0
cycler                  0.11.0
Cython                  3.0.0a9
dask                    2022.2.0
decorator               5.1.1
detectron2              0.6
easymocap               0.2         /home/comp/21482098/workspace/lab/EasyMocap
fonttools               4.28.2
fsspec                  2022.5.0
future-fstrings         1.2.0
google-auth-oauthlib    0.4.6
grpcio                  1.43.0
hydra-core              1.1.1
imageio                 2.19.3
importlib-metadata      4.10.1
ipython                 7.34.0
jedi                    0.18.1
kiwisolver              1.3.2
locket                  1.0.0
Markdown                3.3.6
matplotlib              3.5.2
matplotlib-inline       0.1.3
networkx                2.6.3
numpy                   1.21.6
oauthlib                3.2.0
omegaconf               2.1.1
opencv-python           4.6.0.66
packaging               21.3
parso                   0.8.3
partd                   1.2.0
pexpect                 4.8.0
pickleshare             0.7.5
Pillow                  9.2.0
pip                     22.1.2
prompt-toolkit          3.0.30
protobuf                3.20.1
ptyprocess              0.7.0
pyaml                   21.10.1
pyasn1                  0.4.8
pyasn1-modules          0.2.8
pycocotools             2.0.4
pydot                   1.4.2
pyembree                0.1.6
Pygments                2.12.0
PyOpenGL                3.1.6
pyparsing               3.0.9
python-dateutil         2.8.2
pytorch-fid             0.2.1
PyWavelets              1.3.0
PyYAML                  6.0
regex                   2022.1.18
requests-oauthlib       1.3.1
rsa                     4.8
Rtree                   1.0.0
scikit-image            0.14.2
scipy                   1.7.3
setuptools              63.1.0
setuptools-scm          6.3.2
six                     1.16.0
svgwrite                1.4.2
tensorboard-data-server 0.6.1
tensorboard-plugin-wit  1.8.1
tensorboardX            2.5.1
termcolor               1.1.0
tifffile                2021.11.2
toml                    0.10.2
tomli                   1.2.2
toolz                   0.11.2
torch                   1.8.1+cu101
torchaudio              0.8.1
torchvision             0.9.1+cu101
tqdm                    4.64.0
traitlets               5.3.0
Tree                    0.2.4
trimesh                 3.12.7
typing_extensions       4.3.0
wcwidth                 0.2.5
Werkzeug                2.0.2
wheel                   0.37.1
yacs                    0.1.8
zipp                    3.8.0
